+++ src/main/java/com/example/da/controller/HomeController.java
@@ -1,13 +1,5 @@
 package com.example.da.controller;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/")
-    public String home() {
-        return "index"; // Trả về file index.jsp trong thư mục WEB-INF/views/
-    }
+public interface HomeController {
+    String home();
 }
